rework/moved verification to manageCustomer page update/linked mg-res&mg-cus update/added status to customer data tables update/applied enum change in redactCustomer method rename/customer status enum hidden -> deleted
