+++ src/main/resources/templates/management/verifyCustomer.html
@@ -43,8 +43,8 @@
                 </tr>
                 <tr>
                     <td>
-                        <button type="submit" class="btn btn-dark" th:text="Confirm" onclick="setFormAction('/mg-res/verify/create', 0)"></button>
-                        <button type="submit" class="btn btn-dark" th:text="Back" onclick="setFormAction('/mg-res/verify/back', 0)"></button>
+                        <button type="submit" class="btn btn-dark" th:text="Confirm" onclick="setFormAction('/mg-cus/verify/create', 0)"></button>
+                        <button type="submit" class="btn btn-dark" th:text="Back" onclick="setFormAction('/mg-cus/verify/back', 0)"></button>
                     </td>
                 </tr>
                 </tbody>
