+++ src/main/resources/templates/management/viewCustomer.html
@@ -37,6 +37,16 @@
                     </tr>
 
                     <tr class="table-dark">
+                        <td>Status:</td>
+                    </tr>
+                    <tr>
+                        <td>
+                            <div class="overflow-cell-hs"
+                                 th:text="${customer.getStatus().name().substring(7)}"></div>
+                        </td>
+                    </tr>
+
+                    <tr class="table-dark">
                         <td>Contact:</td>
                     </tr>
                     <tr>
@@ -67,12 +77,44 @@
                             </div>
                         </td>
                     </tr>
+                    </tbody>
+                </table>
+            </div>
+        </th:block>
+
+        <th:block th:if="${customer} != null">
+            <div class="col-sm-9">
+                <table class="table table-borderless table-hover table-fixed"
+                       style="text-align: center; vertical-align: middle;">
+                    <thead class="table-dark">
+                    <tr>
+                        <th scope="row">Option</th>
+                        <th scope="row">Description</th>
+                    </tr>
+                    </thead>
+                    <tbody>
+                    <form class="change_action_form" th:action="@{submit}" th:method="POST">
+                        <input type="hidden" th:if="${customer != null}" th:name="customer"
+                               th:value="${customer.getId()}">
+                        <tr>
+                            <td>
+                                <button type="submit" class="btn btn-dark bg-dark" name="department" th:value="${department}" onclick="setFormAction('/mg-cus/mg-res', 0)">
+                                    <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-person" viewBox="0 0 16 16">
+                                        <path d="M8 8a3 3 0 1 0 0-6 3 3 0 0 0 0 6Zm2-3a2 2 0 1 1-4 0 2 2 0 0 1 4 0Zm4 8c0 1-1 1-1 1H3s-1 0-1-1 1-4 6-4 6 3 6 4Zm-1-.004c-.001-.246-.154-.986-.832-1.664C11.516 10.68 10.289 10 8 10c-2.29 0-3.516.68-4.168 1.332-.678.678-.83 1.418-.832 1.664h10Z"/>
+                                    </svg>
+                                    Reservations
+                                </button>
+                            </td>
+                            <td th:text="'View customer list of reservations'"></td>
+                        </tr>
+                    </form>
+
                     <th:block th:if="${!is_verified}">
-                        <form th:action="@{/mg-res/verify}" th:method="POST">
+                        <form th:action="@{/mg-cus/verify}" th:method="POST">
                             <input type="hidden" th:name="department" th:value="${department}">
                             <tr>
                                 <td style="padding-left:0;">
-                                    <button class="btn btn-dark bg-dark mt-2"
+                                    <button class="btn btn-dark bg-dark"
                                             type="submit" name="customer" th:value="${customer.getId()}">
                                         <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16"
                                              fill="currentColor" class="bi bi-person-check" viewBox="0 0 16 16">
@@ -82,17 +124,18 @@
                                         Verify
                                     </button>
                                 </td>
+                                <td th:text="'Register verification details'"></td>
                             </tr>
                         </form>
                     </th:block>
                     <th:block th:if="${is_verified}">
-                        <form th:action="@{/mg-res/unverify}" th:object="${unverifyConfirmationForm}" th:method="POST">
+                        <form th:action="@{/mg-cus/unverify}" th:object="${unverifyConfirmationForm}" th:method="POST">
                             <input type="hidden" th:name="department" th:value="${department}">
                             <input type="hidden" th:name="customer" th:value="${customer.getId()}">
                             <tr>
                                 <td style="padding-left:0;">
                                     <!--Trigger for modal-->
-                                    <button type="button" class="btn btn-dark bg-dark mt-2" data-bs-toggle="modal"
+                                    <button type="button" class="btn btn-dark bg-dark" data-bs-toggle="modal"
                                             data-bs-target="#unverifyModal">
                                         <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16"
                                              fill="currentColor" class="bi bi-person-dash" viewBox="0 0 16 16">
@@ -132,31 +175,16 @@
                                         </div>
                                     </div>
                                 </td>
+                                <td th:text="'Remove verification details (P-ID, D-ID)'"></td>
                             </tr>
                         </form>
                     </th:block>
-                    </tbody>
-                </table>
-            </div>
-        </th:block>
 
-        <th:block th:if="${customer} != null">
-            <div class="col-sm-9">
-                <table class="table table-borderless table-hover table-fixed"
-                       style="text-align: center; vertical-align: middle;">
-                    <thead class="table-dark">
                     <tr>
-                        <th scope="row">Option</th>
-                        <th scope="row">Description</th>
+                        <td>WIP</td>
+                        <td>WIP</td>
                     </tr>
-                    </thead>
-                    <form th:action="@{submit}" th:method="POST">
-                        <input type="hidden" th:if="${customer != null}" th:name="customer"
-                               th:value="${customer.getId()}">
-                        <tbody>
-                            <td th:text="'W.I.P.'"></td>
-                            <td th:text="'W.I.P.'"></td>
-                        </tbody>
+                    </tbody>
                     </form>
                 </table>
             </div>
